bgpd: Add a bit of debug to give us data about how an update group is formed
The creation of the update group is a tiny bit of a black box and is hard
to figure out by hand if it is correct.  Add a bit of code so that FRR
operators/developers can determine if the update group categorization
makes us happy.

Signed-off-by: Donald Sharp <sharpd@nvidia.com>
Discussed-in: Merge-Request 17371543 7/15, URL: https://code.alibaba-inc.com/SONiC/alibgp/codereview/17371543