fix: peer last reset reason
In community's code, theres a bug in peer_create, that is,
the peer_active(peer) is invoked before the initialization of a peer is done,
and the reason of an inactive peer is always being set as
"No AFI/SAFI activated for peer" due to the check is too early.

Besides, last_reset was not properly inherited.