zebra: fix OOM issue when displaying large route tables in JSON
This is based on FRR community.

When we use command like "show ip route vrf all json",
a root json object would be created, dumped, and finally
freed all together.
However, if theres a large route table,
this json object would occupy a significant amount of memory.
And in some cases, it may lead to daemon crash due to the insufficient memory.
Fix it.

Thus, separating a big json object into small ones by prefix,
so that we can create, display and free them one by one,
and it works well.

In the test based on 61 machine,
the memory for zebra reduces from 321924 KB to 315624 KB,
and for bgpd, from 192992 KB to 186876 KB.

PRLink: https://github.com/FRRouting/frr/pull/16093

Added-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>