bgpd: Add support for tunnel nexthop as bestpath candidate
Currently, if ip nexthop of a path is invalid,
the path would be invalid. But actually we can use te policy, instead.

This commit adds support for tunnel nexthop as bestpath candidate when
ip nexthop is invalid,
by setting configuration &quot;bgp bestpath nexthop-resolved tunnel&quot; on.

This commit is a part of the modification,
others in sonic-cli and sonic-buildimage repository.

Link: https://code.alibaba-inc.com/SONiC/alibgp/codereview/19153776

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>