tests: Fix srv6_vpn use case in multiple threads
Link: https://code.alibaba-inc.com/SONiC/alibgp/codereview/19728743

While running topotests in multiple threads,
the resources such as bandwidth(both for network and IO),
memory, calculation and so on would reduce for each thread.
So if we keep the waiting time as used before,
errors would occur because of the increase of delay time.

Fix this then it can run correctly under 32 threads.
Also fix some other small mistakes.

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>