tests: Fix bgp_link_bw_ip topotest use case for multiple threads
Link: https://code.alibaba-inc.com/SONiC/alibgp/codereview/19732874

While running topotests in multiple threads,
not only the resource issues should be taken
into consideration, but the sequence dependency as well.
Because the order of use cases cannot be guaranteed
in multiple threads.

In the use cases of bgp_link_bw_ip test,
the subsequent validation depends on the previous configuration.
That means, if a later test had run earlier than it should have,
some necessary configuration would be skipped or lost.

Fix it by adding the lost configurations.

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>