tests: Fix topotest use case bgp_conditional_advertisement for multiple threads
Link: https://code.alibaba-inc.com/SONiC/alibgp/codereview/19822465

While running topotests in multiple threads,
the topology state would be maintained within a thread
and be used in the next case.
Therefore, the random order leads to the random state,
and the error would occur.

To break the sequential dependency,
set a function to clean up the topology in each use case and
restore the network to its proper state.
Fix it.

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>