tests: Fix topotest error in srv6_vpn
While we are running the topotests in multiple threads,
and the number is ge 7, an error in test_memory_leak would occur.
It's because test_memory_leak use case would try accessing
the nonexisting "summary.txt" file
though it itself should have been skipped.

summary.txt is generated during RequireVPNRoutes,
but when the number of threads is ge 7,
test_memory_leak will be distributed on its own thread.
As a result, the summary.txt file cannot be generated.

The original access implementation reads the file without
checking its state, so an error occurs in this case.

Fix it by optimizing file access and log as warning.

Link: https://code.alibaba-inc.com/SONiC/alibgp/codereview/19965951

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>