Fix an issue: when isolating the device, the default route advertised to the neighbor could not be correctly withdrawn.
Constructing a default route specifically for a neighbor through a single pass of a route-map can be quite complex.
Here, we consider using a high-priority policy exclusively for isolation purposes, without employing scenarios that match a default route within the high-priority policy.
Therefore, it is determined that when using the high-priority policy, the default route will be withdrawn.

Thus, attention may be needed in the following scenarios:
1.If you wish to still advertise the default route externally during isolation.
2.If you are using the high-priority policy alone and there is no isolation requirement, there will still be an additional withdrawal of the default route.