bgpd: Free mpath in bgp_dest_unlock_node
Besides in bgp_node_destroy, there is another way to free
the bgp_dest and the route_node. The difference between them is the order:
bgp_dest first or route_node first.

In the previous memory optimization for bgp_path_info_mpath,
that is, moving mpath from bgp_path_info to bgp_dest,
we forgot this way, which leads to a memory leak of some bgp_path_info_mpath s.

Let's fix it.

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>