bgpd, zebra: Fix compiling issues for arp2host-backup
Two changes need to be done in previous commits while we
forgot them because we're not using arp2host-backup option.
Compilation would fail currently if we use --enable-arp2host-backup parameter.
Let's fix them.

One is caused by the warning fixing for Werror regarding a
declaration of an only-used var for arp2host-backup as an unused var.
We removed this var earlier to avoid this compiling issue,
but failed to take the exact arp2host-backup mode into consideration.
The corresponding link is:
https://code.alibaba-inc.com/SONiC/alibgp/commit/d8d595651bcb96cd5679e623905950aeca527878
Fix it by move the declaration inside the #ifindex part.

The other is missed in the bgp_node breakup optimization.
The corresponding link is:
https://code.alibaba-inc.com/SONiC/alibgp/commit/82753ea11ac6b92433884e2e4e7b7463d741bc27
Use bd->rn->p to access prefix instead of bd->p to fix.

Find this defect in aone:
https://project.aone.alibaba-inc.com/v2/project/2062098/bug/71164562

Signed-off-by: Yuqing Zhao <galadriel.zyq@alibaba-inc.com>